requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
certifi==2024.7.4
cffi==1.17.0
contourpy==1.2.1
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
devtools==0.12.2
distro==1.9.0
executing==2.0.1
flexcache==0.3
flexparser==0.3.1
fonttools==4.53.1
h11==0.14.0
h3==3.7.7
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
jiter==0.7.0
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mysql-connector-python==9.0.0
numpy==2.0.1
openai==1.54.3
packaging==24.1
parso==0.8.4
pendulum==3.0.0
pexpect==4.9.0
pillow==10.4.0
Pint==0.24.3
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pytz==2024.1
regex==2024.7.24
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
time-machine==2.15.0
timezonefinder==6.5.2
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unit-convert==1.2.0
wcwidth==0.2.13

SQLAlchemy~=2.0.36
alembic~=1.14.0